This PR tries to improve the API lookup calls.

One hotspot is the the `Package.is_vulnerable` property that runs an exists() on a queryset
in a repeated manner.

To avoid this, the approach is to annotate the Package queryset(s) with an `is_vulnerable`
attribute.

The results with a single API call to http://127.0.0.1:8001/api/packages/63 are like this
using pyinstrument and django-toolbar on a laptop:

**Before**
* pyinstrument:
```
Absolute time
Duration: 1.46 seconds
Samples: 1376
CPU time: 1.3 seconds
```
* django-toolbar:
```
Time
Resource 	Value
User CPU time 	3992.892 msec
System CPU time 	69.885 msec
Total CPU time 	4062.777 msec
Elapsed time 	4162.812 msec
Context switches 	13364 voluntary, 52 involuntary

SQL
224.49 ms (1329 queries including 1327 similar and 1299 duplicates ) 
```

The pyinstrument hotspot are calls to:
```
Package.is_vulnerable
vulnerabilities/models.py:647
```

**After**:

* pyinstrument:
```
Absolute time
Duration: 0.288 seconds
Samples: 276
CPU time: 0.263 seconds
```

* django-toolbar:

```
Time
Resource 	Value
User CPU time 	395.396 msec
System CPU time 	20.186 msec
Total CPU time 	415.582 msec
Elapsed time 	439.888 msec
Context switches 	2588 voluntary, 17 involuntary

SQL
83.03 ms (99 queries including 96 similar and 68 duplicates ) 
```


Or an improvement looking like this:

| tool           | old       | new      | improvement 
|------          |--------   |----      |---
| pyinstrument   | 1.46 s    | 0.288 s  | x 5.06
| toolbar        | 4.16 s    | 0.439 s  | x 9.47


```


Reference:https://github.com/nexB/vulnerablecode/issues/1538
Signed-off-by: Philippe Ombredanne <pombredanne@nexb.com>